pyproject.toml
[tool.poetry]
name = "prediction-agent-benchmarker"
version = "0.1.0"
description = ""
authors = ["evangriffiths <egriffiths93@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain-community = "^0.0.16"
pandas = "^2.1.0"
python-dotenv = "^1.0.1"

[tool.poetry.dependencies.evo_researcher]
path = "examples/submodules/evo_researcher"

## Incompatible deps with evo_researcher
# [tool.poetry.dependencies.mech]
# path = "examples/submodules/mech"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
